feat(scrollbar): stage 3
