+++ libs/components/src/lib/components/scrollbar/scrollbar.module.ts
@@ -4,10 +4,16 @@ import { PrizmScrollRefDirective } from './scroll-ref.directive';
 import { PrizmScrollbarComponent } from './scrollbar.component';
 import { PrizmScrollableDirective } from './scrollable.directive';
 import { PrizmScrollControlsModule } from './scroll-controls.module';
+import { PrizmScrollbarRefDirective } from './scrollbar-ref.directive';
 
 @NgModule({
   imports: [CommonModule, PrizmScrollControlsModule],
-  declarations: [PrizmScrollbarComponent, PrizmScrollRefDirective, PrizmScrollableDirective],
+  declarations: [
+    PrizmScrollbarRefDirective,
+    PrizmScrollbarComponent,
+    PrizmScrollRefDirective,
+    PrizmScrollableDirective,
+  ],
   exports: [PrizmScrollbarComponent, PrizmScrollRefDirective, PrizmScrollableDirective],
 })
 export class PrizmScrollbarModule {}
